+++ FluidWarfare.Editor.Windows/Panels/Viewport/NativeHost/Win32/Win32ViewportDefaultProc.cs
@@ -1,0 +1,9 @@
+using System.Runtime.InteropServices;
+
+namespace FluidWarfare.Editor.Windows.Panels.Viewport.NativeHost.Win32;
+
+internal static class Win32ViewportDefaultProc
+{
+    [DllImport("user32.dll", EntryPoint = "DefWindowProcW")]
+    public static extern nint DefWindowProc(nint hwnd, uint msg, nint wParam, nint lParam);
+}